web-app/requirements.txt
Flask==2.2.5
Flask-Login==0.6.2
pymongo==4.7.0
python-dotenv==1.0.0
Werkzeug==2.2.3
opencv-python>=4.5.5
fer>=0.4.0

